server/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jplephem==2.22
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.1.2
pillow==10.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
sgp4==2.23
skyfield==1.49
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
